reorganize extensions
